Added support for glob patterns
Running `a11y` from the command line has glob pattern support but
the grunt plugin was missing this. This is due to the fact that
glob pattern support exists in the cli.js file which is not
exposed as such by the `a11y` module.

So now it should be possible to configure the plugin like this

```js
grunt.initConfig({
  a11y: {
    dev: {
      options: {
        urls: ['**/*.html']
      }
    }
  }
});
```

Unit tests showing this was added too using a black-box (and
somewhat crude) approach. I couldn't find a better way of testing
this for now.